Proposed update in support wording
The previous wording sounded like the workshop was co-organized by these project/institutes which was not the case. This was an IAB workshop and I think we never really have sponsors (except for the physical host venue if applicable). Would the proposed wording work instead? Or @nllz maybe make a different proposal!